+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2022-03-16
+### Added
+- added code for updating time stamps to one second intervals. 
+### Changed
+- Updated slope calculation to remove extra rounding for floating point precision.
+- Updated slope display to add a decimal place.
+### Fixed
+- slope display when the value is 0.
+
 ## [1.8.0] - 2021-05-04
 ### Added
 - added an undo/redo stack